Changed a method in Movie.java
Changed from Double to Integer in updateRating method

This fixes #20

+++ movieRating/core/src/main/java/core/movie.java
@@ -53,7 +53,7 @@ public class Movie {
         this.rating = rating;
     }
 
-    public void updateRating( Double myRating){
+    public void updateRating( int myRating){
         Double sum = ratings.stream().mapToDouble(d -> d).sum();
         this.rating  = sum/ratings.size();  
         register.addMovie(this);
